Don't load Typesafe config yourself
Sentinal shouldn't load Typesafe config itself, since this undermines
the ability to configure it per test, or to use a custom load mechanism
for Typesafe config. It also shouldn't store the config statically.

This converts the config to an Akka extension, and uses the config
loaded by Akka, which gives end users the flexbility to configure it
however they want, including programatically. By making it an Akka
extension, it means the configuration values are only read from the
configuration once per actor system (they are effectively cached for
each use).